requirements.txt
Django==1.8.1
Markdown==2.6.2
Pillow==2.9.0
SQLAlchemy==1.0.4
argparse==1.2.1
boto==2.38.0
dj-database-url==0.3.0
dj-static==0.0.6
django-admin-bootstrapped==2.5.4
django-filter==0.10.0
django-inmemorystorage==0.1.1
django-money==0.7.0
django-postgrespool==0.3.0
django-smart-selects==1.1.1
django-sslify==0.2.7
django-storages==1.1.8
django-toolbelt==0.0.1
djangorestframework==3.1.3
funcsigs==0.4
gunicorn==19.3.0
mock==1.3.0
pbr==1.4.0
psycopg2==2.6
py-moneyed==0.5.0
six==1.9.0
static3==0.6.1
whitenoise==1.0.6
wsgiref==0.1.2
